Fix erfinv and swapaxes

+++ oneflow/user/kernels/erfinv_kernel.cpp
@@ -52,14 +52,16 @@ class CpuErfinvKernel final : public user_op::OpKernel {
         dem = (d[1] * z + d[0]) * z + static_cast<T>(1.0);
         y_ptr[i] = std::copysign(num, x) / dem;
       }
-      y_ptr[i] = y_ptr[i]
-                 - (std::erf(y_ptr[i]) - x)
-                       / ((static_cast<T>(2.0) / static_cast<T>(std::sqrt(M_PI)))
-                          * std::exp(-y_ptr[i] * y_ptr[i]));
-      y_ptr[i] = y_ptr[i]
-                 - (std::erf(y_ptr[i]) - x)
-                       / ((static_cast<T>(2.0) / static_cast<T>(std::sqrt(M_PI)))
-                          * std::exp(-y_ptr[i] * y_ptr[i]));
+      y_ptr[i] =
+          y_ptr[i]
+          - (std::erf(y_ptr[i]) - x)
+                / ((static_cast<T>(2.0) / static_cast<T>(std::sqrt(static_cast<double>(M_PI))))
+                   * std::exp(-y_ptr[i] * y_ptr[i]));
+      y_ptr[i] =
+          y_ptr[i]
+          - (std::erf(y_ptr[i]) - x)
+                / ((static_cast<T>(2.0) / static_cast<T>(std::sqrt(static_cast<double>(M_PI))))
+                   * std::exp(-y_ptr[i] * y_ptr[i]));
     }
   }
